+++ ProjectGroenBos/Schoonmaak en Onderhoud/OffertesOverzicht.aspx.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.UI;
+using System.Web.UI.WebControls;
+
+namespace ProjectGroenBos.Schoonmaak_en_Onderhoud
+{
+    public partial class OffertesOverzicht : System.Web.UI.Page
+    {
+        protected void Page_Load(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
