--- Schoonmaak en Onderhoud/PlanningsOverzicht.aspx
@@ -1,15 +1,0 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="PlanningsOverzicht.aspx.cs" Inherits="ProjectGroenBos.Schoonmaak_en_Onderhoud.PlanningsOverzicht" %>
-
-<!DOCTYPE html>
-
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head runat="server">
-    <title></title>
-</head>
-<body>
-    <form id="form1" runat="server">
-        <div>
-        </div>
-    </form>
-</body>
-</html>
